+++ src/instance/constructor.js
@@ -31,7 +31,7 @@ export default function ScrollReveal(options = {}) {
 		return new ScrollReveal(options)
 	}
 
-	if (!ScrollReveal.isSupported()) {
+	if (!transformSupported() || !transitionSupported()) {
 		logger.call(this, 'Instantiation aborted.', 'This browser is not supported.')
 		return noop
 	}
@@ -105,12 +105,6 @@ export default function ScrollReveal(options = {}) {
 
 	return _instance ? _instance : (_instance = this)
 }
-
-/**
- * Static members are available immediately during instantiation,
- * so debugging and browser support details are handled here.
- */
-ScrollReveal.isSupported = () => transformSupported() && transitionSupported()
 
 Object.defineProperty(ScrollReveal, 'debug', {
 	get: () => _debug || false,